backend/requirements1.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.10.4
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.6
click==8.3.1
cryptography==42.0.8
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.24
frozenlist==1.8.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.0
idna==3.11
Jinja2==3.1.4
limits==5.8.0
loguru==0.7.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
orjson==3.11.7
packaging==26.0
passlib==1.7.4
pillow==10.4.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.20.0
pyparsing==3.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.3
reportlab==4.2.2
requests==2.33.0
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
